Move permission flow into latchkey/, make routes thin dispatchers
The previous PermissionGrantHandler was generic-named but latchkey-specific
in every detail: it called Latchkey.services_info / auth_browser, wrote
detent-format permissions.json, and stamped LATCHKEY_PERMISSION on the
response event. It now lives at latchkey/permissions.py as
LatchkeyPermissionGrantHandler and owns:

  * the latchkey services catalog (no more app.state.latchkey_services_catalog),
  * the dialog rendering (render_request_page),
  * the grant/deny HTTP entry points (apply_grant_request /
    apply_deny_request),
  * a small display_name_for_event helper used by the requests panel.

Pure grant() and deny() methods are kept (and now return the freshly-
appended response event so the HTTP layer can mirror the same instance
into the in-memory inbox without minting a new event_id).

The route handlers in app.py become type-dispatchers: they authenticate,
look up the request event, and delegate to the type-specific handler:

  * _handle_request_page: dispatches to handler.render_request_page
    for LatchkeyPermissionRequestEvent (and the existing sharing path
    for SharingRequestEvent).
  * _handle_permission_grant / _handle_permission_deny: dispatch to
    handler.apply_grant_request / apply_deny_request. No latchkey-specific
    code remains in app.py.

Also rename PermissionFlowError to LatchkeyPermissionFlowError now that
it lives next to the latchkey handler.

The asyncio-import ratchet bumps from 1 to 2 because the handler uses
run_in_executor to keep blocking grant work off the event loop.
